Forward request timeouts
Consider the following scenario: a user forgets to turn a robot on, tries to
connect to it, realizes the mistake, turns the robot on, and reattempts the
connection within 60 seconds. The initial, failed connection attempt will have
generated a throwaway disconnect request to ensure that if the robot does
eventually receive the initial connect request, it will be closed.

At the application level, the disposition of that disconnect request is
ignored. However, the daemon must create a forwarding request for that
disconnect request. The forwarding request has a much longer request timeout
interval, so as not to artificially constrain the application's timeout
interval. When the forwarding request timeout comes due, the daemon would
erroneously close the proxy which, in our scenario, had been saved from death
by the user's retry attempt within sixty seconds.

The answer: if a forwarding request times out, forward the timeout. The
application/library code already ignores this particular case.